riscv: sifive: Rename sifive_prci.{c,h} to sifive_e_prci.{c,h}
Current SiFive PRCI model only works with sifive_e machine, as it
only emulates registers or PRCI block in the FE310 SoC.

Rename the file name to make it clear that it is for sifive_e.
This also prefix "sifive_e"/"SIFIVE_E" for all macros, variables
and functions.

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>
Reviewed-by: Chih-Min Chao <chihmin.chao@sifive.com>
Reviewed-by: Alistair Francis <alistair.francis@wdc.com>

Series-changes: 4
- prefix all macros/variables/functions with SIFIVE_E/sifive_e
  in the sifive_e_prci driver