riscv: hart: Add a "hartid-base" property to RISC-V hart array
At present each hart's hartid in a RISC-V hart array is assigned
the same value of its index in the hart array. But for a system
that has multiple hart arrays, this is not the case any more.

Add a new "hartid-base" property so that hartid number can be
assigned based on the property value.

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>
Reviewed-by: Alistair Francis <alistair.francis@wdc.com>

Series-changes: 7
- use `s->hartid_base + idx` directly

Series-changes: 6
- use s->hartid_base directly, instead of an extra variable

Series-changes: 4
- new patch to add a "hartid-base" property to RISC-V hart array