hw/ssi: imx_spi: Round up the burst length to be multiple of 8
Current implementation of the imx spi controller expects the burst
length to be multiple of 8, which is the most common use case.

In case the burst length is not what we expect, log it to give user
a chance to notice it, and round it up to be multiple of 8.

Signed-off-by: Bin Meng <bin.meng@windriver.com>

Series-changes: 9
- Do the LOG_UNIMP when the unsupported burst length value is written,
  rather than where it is used.
- Squash the 2 LOG_UNIMP warnings down into one line

Series-changes: 5
- round up the burst length to be multiple of 8

Series-changes: 4
- s/normal/common/ in the commit message
- log the burst length value in the log message

Series-changes: 3
- new patch: log unimplemented burst length